requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.3
pydantic>=2.5.3
google-generativeai>=0.8.0
